Add interview pack and summary features, enhance analytics and UI
- Add interview pack generation: tailored questions based on job description, CV, and interview type
- Add interview summary generation: AI-powered summaries with overall scores, strengths, and concerns
- Add InterviewDetail page: dedicated page for individual interviews with notes and scoring
- Enhance CandidateDetail: show interview summaries in interview cards, navigate to interview detail page
- Update RoleDetails: sort by combined score (CV + interview), display interview results
- Add status history tracking: automatically log interview bookings and completions
- Enhance PDF summary: include interview results, combined scores, and comprehensive candidate profiles
- Remove hardcoding from Analytics: calculate real metrics from actual data
- Simplify Status page: remove API references, focus on AI Scoring and Site status
- Update Navbar: full-width layout with logo on left edge, navigation on right edge
- Remove API Keys tab from Settings
- Remove Documentation link from Footer
- Various UI/UX improvements and bug fixes

+++ apps/web/DATABASE_SCHEMA.sql
@@ -239,11 +239,11 @@ CREATE TABLE IF NOT EXISTS public.candidates (
   cv_file_url TEXT,
   cv_parsed_data JSONB, -- Full parsed CV data from ParseScore API
 
-  -- Scoring & Matching
-  overall_score INTEGER CHECK (overall_score >= 0 AND overall_score <= 100),
-  skills_match_score INTEGER CHECK (skills_match_score >= 0 AND skills_match_score <= 100),
-  experience_score INTEGER CHECK (experience_score >= 0 AND experience_score <= 100),
-  education_score INTEGER CHECK (education_score >= 0 AND education_score <= 100),
+  -- Scoring & Matching (NUMERIC for decimal precision)
+  overall_score NUMERIC(5,2) CHECK (overall_score >= 0 AND overall_score <= 100),
+  skills_match_score NUMERIC(5,2) CHECK (skills_match_score >= 0 AND skills_match_score <= 100),
+  experience_score NUMERIC(5,2) CHECK (experience_score >= 0 AND experience_score <= 100),
+  education_score NUMERIC(5,2) CHECK (education_score >= 0 AND education_score <= 100),
 
   -- Extracted Information
   skills JSONB DEFAULT '[]'::jsonb, -- Array of skills
